Derive `std::fmt::Debug` for the `domain` types
In roder to facilitate debugging, derive the standard implementation of the
`std::fmt::Debug` trait for the `domain` types. Add an according
`debug!` in `read_block` for pretty-printing each block.

Pretty-printing a `Block` also prints all its contained `Transaction`s and their
`Input`s and `Output`s which currently results in a lot of debug-output.